Port original git-sync script from bash
The goal is to get the local repository fully in-sync with all remotes, including fast-forwarding the local copy of branches like "main". Additionally, if there are different origin and upstream remotes, keep origin's copy of main tracking upstream as well, as this avoids occasional mistakes.

This implementation assumes the only local branch that needs to track its remote upstream is called "main", and that the remote origin copy of main needs to be kept up-to-date.